backend/requirements.txt
# Core Framework
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0
python-dotenv>=1.0,<2.0

# Database & Storage
supabase>=2.4,<2.5
postgrest>=0.16,<0.17

# HTTP Client & AI Providers
httpx>=0.27,<1.0
groq>=0.11,<1.0

# Document Processing
pypdf>=6.8.0,<7.0
Pillow>=12.1.1,<13.0

# Security & Utilities
cryptography>=46.0.5,<47.0
pydantic>=2.5,<3.0
email-validator>=2.1,<3.0

# Rate Limiting & Caching
slowapi>=0.1.9,<1.0

# System Monitoring
psutil>=5.9,<8.0

# Development & Testing (optional)
pytest>=8.3,<9.0
pytest-asyncio>=0.21,<2.0
